install: do not overwrite config if it exists

+++ install.sh
@@ -1,3 +1,5 @@
 set -x
 cp pidgin-pipe-status.pl ~/.purple/plugins/
-cp pidgin-pipe-status-config.properties ~/.purple/plugins/
+if [ ! -e ~/.purple/plugins/pidgin-pipe-status-config.properties ]; then
+  cp pidgin-pipe-status-config.properties ~/.purple/plugins/
+fi
